Bindings for new location
Remove bindings and test from old location. This erases the history for
the bindings and the test; but they are sufficiently slight that it's not
a great loss.
