+++ bindings/pydrake/multibody/_parsing_extra.py
@@ -1,0 +1,2 @@
+# For backwards compatibility.
+from pydrake.common.parsing import PackageMap
