+++ bindings/pydrake/multibody/test/parsing_test.py
@@ -13,7 +13,7 @@ from pydrake.multibody.parsing import (
     ModelDirective,
     ModelDirectives,
     ModelInstanceInfo,
-    PackageMap,
+    PackageMap,  # Confirm backwards compatibility.
     Parser,
     ProcessModelDirectives,
 )
@@ -24,6 +24,7 @@ import re
 import unittest
 
 from pydrake.common import FindResourceOrThrow
+from pydrake.common.parsing import PackageMap
 from pydrake.common.test_utilities.deprecation import catch_drake_warnings
 from pydrake.geometry import SceneGraph
 from pydrake.multibody.tree import (
@@ -35,35 +36,6 @@ from pydrake.multibody.plant import (
 
 
 class TestParsing(unittest.TestCase):
-
-    def test_package_map(self):
-        # Simple coverage test for constructors.
-        dut = PackageMap()
-        self.assertEqual(dut.size(), 2)
-        PackageMap(other=dut)
-        copy.copy(dut)
-
-        dut = PackageMap.MakeEmpty()
-        dut2 = PackageMap.MakeEmpty()
-        tmpdir = os.environ.get('TEST_TMPDIR')
-
-        # Simple coverage test for Add, AddMap, Contains, size,
-        # GetPackageNames, GetPath, AddPackageXml, Remove.
-        dut.Add(package_name="root", package_path=tmpdir)
-        dut2.Add(package_name="root", package_path=tmpdir)
-        dut.AddMap(dut2)
-        self.assertTrue(dut.Contains(package_name="root"))
-        self.assertEqual(dut.size(), 1)
-        self.assertEqual(dut.GetPackageNames(), ["root"])
-        self.assertEqual(dut.GetPath(package_name="root"), tmpdir)
-        dut.AddPackageXml(filename=FindResourceOrThrow(
-            "drake/multibody/parsing/test/box_package/package.xml"))
-        dut2.Remove(package_name="root")
-        self.assertEqual(dut2.size(), 0)
-
-        # Simple coverage test for folder and environment.
-        dut.PopulateFromEnvironment(environment_variable='TEST_TMPDIR')
-        dut.PopulateFromFolder(path=tmpdir)
 
     def test_package_map_remote_params(self):
         dut = PackageMap.RemoteParams(
